[feat] POM-008-sequence : Ajout de la liste de timer
