+++ frontend/src/components/timer/Timer.jsx
@@ -1,12 +1,5 @@
 import { useEffect, useState } from "react";
-
-function formatMS(totalSeconds) {
-  const minute = Math.floor(totalSeconds / 60);
-  const second = totalSeconds % 60;
-  const mm = String(minute).padStart(2, "0");
-  const ss = String(second).padStart(2, "0");
-  return `${mm}:${ss}`;
-}
+import { formatMS } from "../../utils/formatTime";
 
 export default function PomodoroTimer({ selectedTime, isStart }) {
   const [timer, setTimer] = useState(selectedTime);
